chore(gittag) factorize plugin a separated package
